d/b/buildcheck.py, d/rules.real: Run buildcheck.py in setup as well
I want to restore some checks on the kernel config.  These can be done
in the setup target, before the full build, and it is preferable to do
that rather than waste time on building with an invalid config.

(cherry picked from commit 8f96ed91418980dcfbe1e71887f726f44030422a)