Generate kernel ABI name suffix automatically if not configured
When we sign modules with a unique ephemeral key, each upload needs to
have a distinct kernel ABI name even if the changes don't otherwise
change the kernel module ABI.

In trixie the kernel version string format was changed significantly
to support this.  This should probably not be done for bookworm and
earlier releases.

So generate the ABI name automatically from the number of changelog
entries, like we do in trixie onward, but:

- Keep the current format: <major>.<minor>.0-<serial>
- Use a base serial number of 38 for the latest version 6.1.147-1
  and add 1 for each subsequent bookworm version
- For backports:
  - Keep the current format: <major>.<minor>.0-0.deb<release>.<serial>
  - Don't bother to add configuration for this, but hardcode the
    mapping for all releases in LTS and ELTS
  - Add a further dotted component when there are multiple backports
    from a single bookworm version to the same release

Retain support for configuring the abiname through
debian/config/defines, but don't bother supporting per-architecture
overrides.