d/b/buildcheck.py: Check config of kernel to be signed
We must only sign kernel images that honour the Secure Boot setting.
gencontrol_signed.py used to check the kernel config file to ensure
that the required options (CONFIG_EFI_STUB and
CONFIG_LOCK_DOWN_IN_EFI_SECURE_BOOT) were enabled.  This is currently
done in gencontrol_signed.py.

We must also list any certificates trusted by the kernel in files.json
in the signing template.  We used to generate this from the
CONFIG_SYSTEM_TRUSTED_KEYS variable in kernel config, but that is now
set to an empty string and gencontrol.py assumes that is the case.
We can't check this in gencontrol_signed.py, because this variable
is filtered out of the installed .config.

So check all of these in buildcheck.py, in the setup phase.

This is roughly a backport of commit
620854e18896ca12be194c5316d8e9f04c1f0aad.