d/salsa-ci.yml: Fix source package name in $BUILD_COMMAND
In the build job, the .dsc filename in $BUILD_COMMAND is currently set
assuming the source package name is "linux".  This is not true for
this backport.  Replace "linux" with "${source}".

Similarly, in the build-signed job, the .dsc filename pattern is set
assuming the source package name is "linux-signed-<arch>", whereas
for this backport it is "linux-signed-6.1-<arch>".  Add a wildcard
to allow for this.